==> notebook/1_data_cleaning.ipynb <==
%pip install pandas
import pandas as pd
%pip install openpyxl
##Load dataset
%pip install openpyxl 
file_path = "C:/Users/NPG/Desktop/maheswar/customer segmentation +clv prediction/data/raw/online_retail_II.xlsx" ## Adjust path as needed
sheet_name = "Year 2010-2011"

print("Loading dataset...")
df = pd.read_excel(file_path, sheet_name=sheet_name, engine='openpyxl')
print(f"Raw data shape: {df.shape}")

# Remove cancellations
df = df[~df['Invoice'].astype(str).str.startswith('C')]
print(f"After removing cancellations: {df.shape}")

# Remove negative or zero Quantity
df = df[df['Quantity'] > 0]
print(f"After removing negative/zero quantities: {df.shape}")

# Drop rows with missing Customer ID
df = df.dropna(subset=['Customer ID'])
print(f"After dropping missing Customer ID: {df.shape}")

# Remove duplicates
df = df.drop_duplicates()
print(f"After removing duplicates: {df.shape}")

# Check missing values in all columns
print("Missing values per column:")
print(df.isnull().sum())
# Remove rows where Price == 0
df = df[df['Price'] > 0]

print(f"New dataset shape after removing rows with Price == 0: {df.shape}")

# Save cleaned dataset
output_path = "data/new_clean/cleaned_online_retail.csv"
df.to_csv(output_path, index=False)
print(f"Cleaned data saved to: {output_path}")
print(df[df['Price'] == 0].shape[0])
